+++ src/Stack.hs
@@ -1,32 +1,35 @@
 module Stack (
   Stack,
+  Box (..),
   push,
   pop,
   popN
 ) where
 
 import Function (Function)
+import Box (Box (..))
 
--- A Stack is simply a list of Functions
-type Stack = [Function]
+-- A Stack is a list of Boxes of Functions
+type Stack = [Box Function]
 
--- Push a Function to a Stack, returning a new Stack
-push :: Function -> Stack -> Stack
+-- Push a Box to a Stack, returning a new Stack
+push :: Box Function -> Stack -> Stack
 push = (:)
 
--- Pop the top Function from a Stack, returning the Function and a
--- new Stack with it missing
--- If the Stack is empty, use the identity function
-pop :: Stack -> (Function, Stack)
-pop (f : fs) = (f, fs)
-pop [] = (mempty, [])
+-- Pop the top Box from a Stack, returning the Box and a new Stack
+-- with it missing
+-- If the Stack is empty, use a Box containing the identity function
+pop :: Stack -> (Box Function, Stack)
+pop (b : bs) = (b, bs)
+pop [] = (Single mempty, [])
 
--- Pop n Functions from a Stack, returning a list of Functions and a
--- new Stack with them missing
--- The Functions are listed with the top of the Stack at the end of
+-- Pop n Boxes from a Stack, returning a list of Boxes and a new Stack
+-- with them missing
+-- The Boxes are returned with the top of the Stack at the end of
 -- the list, in keeping with the order in which they were (probably)
 -- pushed: a program like A B over should give A over B, not B over A
 -- If the Stack does not contain enough elements, fill in with the
 -- identity function
-popN :: Int -> Stack -> ([Function], Stack)
-popN n s = (reverse $ take n $ s ++ cycle [mempty], drop n s)
+popN :: Int -> Stack -> ([Box Function], Stack)
+popN n s = (reverse $ take n s', drop n s)
+  where s' = s ++ cycle [Single mempty]